+++ src/codegen/translator_factory.cpp
@@ -24,7 +24,11 @@
 #include "codegen/order_by_translator.h"
 #include "codegen/table_scan_translator.h"
 #include "codegen/tuple_value_translator.h"
+#include "expression/comparison_expression.h"
+#include "expression/conjunction_expression.h"
 #include "expression/constant_value_expression.h"
+#include "expression/operator_expression.h"
+#include "expression/tuple_value_expression.h"
 #include "planner/aggregate_plan.h"
 #include "planner/hash_join_plan.h"
 #include "planner/order_by_plan.h"
