+++ src/include/codegen/comparison_translator.h
@@ -14,9 +14,13 @@
 
 #include "codegen/compilation_context.h"
 #include "codegen/expression_translator.h"
-#include "expression/comparison_expression.h"
 
 namespace peloton {
+
+namespace expression {
+class ComparisonExpression;
+}  // namespace expression
+
 namespace codegen {
 
 //===----------------------------------------------------------------------===//
@@ -26,40 +30,11 @@ class ComparisonTranslator : public ExpressionTranslator {
  public:
   // Constructor
   ComparisonTranslator(const expression::ComparisonExpression &comparison,
-                       CompilationContext &context)
-      : ExpressionTranslator(comparison, context) {
-    PL_ASSERT(comparison.GetChildrenSize() == 2);
-  }
+                       CompilationContext &context);
 
   // Produce the result of performing the comparison of left and right values
   codegen::Value DeriveValue(CodeGen &codegen,
-                             RowBatch::Row &row) const override {
-    const auto &comparison =
-        GetExpressionAs<expression::ComparisonExpression>();
-
-    codegen::Value left = row.DeriveValue(codegen, *comparison.GetChild(0));
-    codegen::Value right = row.DeriveValue(codegen, *comparison.GetChild(1));
-
-    switch (comparison.GetExpressionType()) {
-      case ExpressionType::COMPARE_EQUAL:
-        return left.CompareEq(codegen, right);
-      case ExpressionType::COMPARE_NOTEQUAL:
-        return left.CompareNe(codegen, right);
-      case ExpressionType::COMPARE_LESSTHAN:
-        return left.CompareLt(codegen, right);
-      case ExpressionType::COMPARE_LESSTHANOREQUALTO:
-        return left.CompareLte(codegen, right);
-      case ExpressionType::COMPARE_GREATERTHAN:
-        return left.CompareGt(codegen, right);
-      case ExpressionType::COMPARE_GREATERTHANOREQUALTO:
-        return left.CompareGte(codegen, right);
-      default: {
-        throw Exception{
-            "Invalid expression type for translation " +
-            ExpressionTypeToString(comparison.GetExpressionType())};
-      }
-    }
-  }
+                             RowBatch::Row &row) const override;
 };
 
 }  // namespace codegen
