+++ src/include/codegen/conjunction_translator.h
@@ -14,9 +14,13 @@
 
 #include "codegen/compilation_context.h"
 #include "codegen/expression_translator.h"
-#include "expression/conjunction_expression.h"
 
 namespace peloton {
+
+namespace expression {
+class ConjunctionExpression;
+}  // namespace expression
+
 namespace codegen {
 
 //===----------------------------------------------------------------------===//
@@ -26,30 +30,11 @@ class ConjunctionTranslator : public ExpressionTranslator {
  public:
   // Constructor
   ConjunctionTranslator(const expression::ConjunctionExpression &conjunction,
-                        CompilationContext &context)
-      : ExpressionTranslator(conjunction, context) {
-    PL_ASSERT(conjunction.GetChildrenSize() == 2);
-  }
+                        CompilationContext &context);
 
   // Produce the value that is the result of codegening the expression
   codegen::Value DeriveValue(CodeGen &codegen,
-                             RowBatch::Row &row) const override {
-    const auto &conjunction =
-        GetExpressionAs<expression::ConjunctionExpression>();
-    codegen::Value left = row.DeriveValue(codegen, *conjunction.GetChild(0));
-    codegen::Value right = row.DeriveValue(codegen, *conjunction.GetChild(1));
-
-    switch (conjunction.GetExpressionType()) {
-      case ExpressionType::CONJUNCTION_AND:
-        return left.LogicalAnd(codegen, right);
-      case ExpressionType::CONJUNCTION_OR:
-        return left.LogicalOr(codegen, right);
-      default:
-        throw Exception{
-            "Received a non-conjunction expression type: " +
-            ExpressionTypeToString(conjunction.GetExpressionType())};
-    }
-  }
+                             RowBatch::Row &row) const override;
 };
 
 }  // namespace codegen
